+++ addons/gut/awaiter.gd
@@ -196,6 +196,6 @@ func wait_while(predicate_function: Callable, max_time, time_between_calls:=0.0,
 
 
 func is_waiting():
-	return _wait_time != 0.0 || \
-		_wait_physics_frames != 0 || \
+	return _wait_time != 0.0 or \
+		_wait_physics_frames != 0 or \
 		_wait_process_frames != 0