+++ addons/gut/doubler.gd
@@ -195,14 +195,14 @@ func _create_double(parsed, strategy, override_path, partial):
 	for method in parsed.get_local_methods():
 		if(_is_method_eligible_for_doubling(parsed, method)):
 			included_methods.append(method.meta.name)
-			dbl_src += _get_func_text(method.meta)
+			dbl_src += _method_maker.get_function_text(method)
 
 	if(strategy == GutUtils.DOUBLE_STRATEGY.INCLUDE_NATIVE):
 		for method in parsed.get_super_methods():
 			if(_is_method_eligible_for_doubling(parsed, method)):
 				included_methods.append(method.meta.name)
 				_stub_to_call_super(parsed, method.meta.name)
-				dbl_src += _get_func_text(method.meta)
+				dbl_src += _method_maker.get_function_text(method)
 
 	var base_script = _get_base_script_text(parsed, override_path, partial, included_methods)
 	dbl_src = base_script + "\n\n" + dbl_src
@@ -238,7 +238,7 @@ func _create_singleton_double(singleton, is_partial):
 	var DblClass = GutUtils.create_script_from_source(dbl_src)
 	if(_stubber != null):
 		for key in parsed.methods_by_name:
-			var meta = parsed.methods_by_name[key]
+			var meta = parsed.methods_by_name[key].meta
 			if(meta != {} and !meta.flags & METHOD_FLAG_VARARG):
 				_stubber.stub_defaults_from_meta(singleton, meta)
 
@@ -248,7 +248,7 @@ func _create_singleton_double(singleton, is_partial):
 func _stub_method_default_values(parsed):
 	for method in parsed.get_local_methods():
 		if(method.is_eligible_for_doubling() and !_ignored_methods.has(parsed.resource, method.meta.name)):
-			_stubber.stub_defaults_from_meta(parsed.script_path, method.meta)
+			_stubber.stub_defaults_from_meta(parsed.resource, method.meta)
 
 
 func _double_scene_and_script(scene, strategy, partial):
@@ -279,10 +279,6 @@ func _get_inst_id_ref_str(inst):
 	if(inst):
 		ref_str = str('instance_from_id(', inst.get_instance_id(),')')
 	return ref_str
-
-
-func _get_func_text(method_hash):
-	return _method_maker.get_function_text(method_hash) + "\n"
 
 
 func _parse_script(obj):