+++ addons/gut/gui/ShortcutButton.gd
@@ -124,7 +124,7 @@ func get_input_event():
 	return _source_event
 
 func set_shortcut(sc):
-	if(sc == null or sc.events == null || sc.events.size() <= 0):
+	if(sc == null or sc.events == null or sc.events.size() <= 0):
 		clear_shortcut()
 	else:
 		_source_event = sc.events[0]
@@ -144,8 +144,8 @@ func disable_set(should):
 
 func disable_clear(should):
 	_ctrls.clear_button.disabled = should
-	
-	
+
+
 func cancel():
 	if(_editing):
 		_edit_mode(false)