+++ addons/gut/script_parser.gd
@@ -8,12 +8,10 @@ const BLACKLIST = [
 
 
 # ------------------------------------------------------------------------------
-# Combins the meta for the method with additional information.
-# * flag for whether the method is local
-# * adds a 'default' property to all parameters that can be easily checked per
-#   parameter
+# Combines the meta for the method with additional information.
 # ------------------------------------------------------------------------------
 class ParsedMethod:
+
 	const NO_DEFAULT = '__no__default__'
 
 	var _meta = {}
@@ -22,7 +20,8 @@ class ParsedMethod:
 		set(val): return;
 
 	var is_local = false
-	var _parameters = []
+	var args = []
+	var return_type_text = 'void'
 
 	func _init(metadata):
 		_meta = metadata
@@ -35,7 +34,22 @@ class ParsedMethod:
 				arg['default'] = _meta.default_args[start_default - i]
 			else:
 				arg['default'] = NO_DEFAULT
-			_parameters.append(arg)
+			args.append(arg)
+
+		return_type_text = _get_return_type(metadata)
+
+	func _get_return_type(meta):
+		var r_meta = meta["return"]
+		var return_keyword = GutConstants.TYPE_KEYWORDS[r_meta.type]
+
+		if(r_meta.type != 0):
+			return_keyword = return_keyword
+		elif(r_meta.usage & PROPERTY_USAGE_NIL_IS_VARIANT != 0):
+			return_keyword = 'Variant'
+		else:
+			return_keyword = 'void'
+
+		return return_keyword
 
 
 	func is_eligible_for_doubling():
@@ -275,9 +289,12 @@ class ParsedScript:
 		return text
 
 
+
+
 # ------------------------------------------------------------------------------
 # ------------------------------------------------------------------------------
 var scripts = {}
+
 
 func _get_instance_id(thing):
 	var inst_id = null