+++ addons/gut/utils.gd
@@ -2,8 +2,6 @@
 class_name GutUtils
 extends Object
 
-const GUT_METADATA = '__gutdbl'
-
 # Note, these cannot change since places are checking for TYPE_INT to determine
 # how to process parameters.
 enum DOUBLE_STRATEGY{
@@ -15,6 +13,13 @@ enum DIFF {
 	DEEP,
 	SIMPLE
 }
+
+enum TREAT_AS {
+	NOTHING,
+	FAILURE,
+}
+
+const GUT_METADATA = '__gutdbl'
 
 const TEST_STATUSES = {
 	NO_ASSERTS = 'no asserts',
@@ -37,17 +42,19 @@ const NOTHING := '__NOTHING__'
 const NO_TEST := 'NONE'
 const GUT_ERROR_TYPE = 999
 
-enum TREAT_AS {
-	NOTHING,
-	FAILURE,
-}
 
-static var class_ref_by_name = {} :
+static var _class_ref_by_name := {}
+static var class_ref_by_name := {} :
 	get():
-		if(class_ref_by_name == {}):
-			class_ref_by_name = _create_class_dictionary()
-		return class_ref_by_name;
-
+		if(_class_ref_by_class == {}):
+			_create_class_dictionaries()
+		return _class_ref_by_name
+static var _class_ref_by_class := {}
+static var class_ref_by_class := {} :
+	get():
+		if(_class_ref_by_class == {}):
+			_create_class_dictionaries()
+		return _class_ref_by_class
 
 
 ## This dictionary defaults to all the native classes that we cannot call new
@@ -204,7 +211,7 @@ static var gut_fonts = GutFonts.new()
 static var avail_fonts = gut_fonts.get_font_names()
 
 static var version_numbers = VersionNumbers.new(
-	'9.6.0' # gut_versrion (source of truth)
+	'9.7.0' # gut_versrion (source of truth)
 )
 
 
@@ -250,17 +257,25 @@ static func get_error_tracker():
 #
 # This is lazy loaded into class_ref_by_name, it's only needed when finding
 # stubs.
-static func _create_class_dictionary():
+static func _create_class_dictionaries():
 	var text = "var all_the_classes: Dictionary = {\n"
 	var black_list = [
+		"IPUnix",
+		"GodotNavigationServer2D",
+		"NativeMenuMacOS",
 	]
 	for classname in ClassDB.get_class_list():
-		if(!black_list.has(classname) and (ClassDB.can_instantiate(classname) or GodotSingletons.names.has(classname))):
+		if(!black_list.has(classname) and (ClassDB.can_instantiate(classname) or \
+		 	GodotSingletons.names.has(classname))):
 			text += str('"', classname, '": ', classname, ", \n")
 
 	text += "}"
 	var inst =  GutUtils.create_script_from_source(text, 'res://dynamically_generated/class_dictionary.gd').new()
-	return inst.all_the_classes
+
+	_class_ref_by_name = inst.all_the_classes
+	for key in inst.all_the_classes:
+		_class_ref_by_class[inst.all_the_classes[key]] = key
+
 
 
 
@@ -643,7 +658,12 @@ static func find_method_meta(methods, method_name):
 
 
 static func get_method_meta(object, method_name):
-	return find_method_meta(object.get_method_list(), method_name)
+	if(object is GDScript):
+		return find_method_meta(object.get_script_method_list(), method_name)
+	elif(is_native_class(object)):
+		return find_method_meta(ClassDB.class_get_method_list(class_ref_by_class[object]), method_name)
+	else:
+		return find_method_meta(object.get_method_list(), method_name)
 
 
 static func is_singleton(thing):