+++ test/integration/test_doubler_and_stubber.gd
@@ -29,7 +29,6 @@ class BaseTest:
 
 
 
-
 class TestTheBasics:
 	extends BaseTest
 
@@ -82,22 +81,25 @@ class TestTheBasics:
 		assert_eq(d2.get_value(), 5, 'other instantiate gets class value')
 
 	func test_stubbed_methods_send_parameters_in_callback():
-		var sp = StubParams.new(DOUBLE_ME_PATH, 'has_one_param')
-		sp.to_return(10).when_passed(1)
+		var sp = StubParams.new(DOUBLE_ME_PATH, 'return_string_plus_a')
+		sp.to_return('foo').when_passed('a')
+		gr.stubber.add_stub(sp)
+		sp = StubParams.new(DOUBLE_ME_PATH, 'return_string_plus_a')
+		sp.to_return('not_stubbed')
 		gr.stubber.add_stub(sp)
 		var d = autofree(gr.doubler.double(DoubleMe).new())
-		assert_eq(d.has_one_param(1), 10)
-		assert_eq(d.has_one_param('asdf'), null)
+		assert_eq(d.return_string_plus_a('a'), 'foo')
+		assert_eq(d.return_string_plus_a('asdf'), 'not_stubbed')
 
 	func test_stub_with_nothing_works_with_parameters():
-		var sp1 = StubParams.new(DOUBLE_ME_PATH, 'has_one_param').to_return(5)
-		var sp2 = StubParams.new(DOUBLE_ME_PATH, 'has_one_param')
-		sp2.to_return(10).when_passed(1)
+		var sp1 = StubParams.new(DOUBLE_ME_PATH, 'return_string_plus_a').to_return('5')
+		var sp2 = StubParams.new(DOUBLE_ME_PATH, 'return_string_plus_a')
+		sp2.to_return('10').when_passed('1')
 		gr.stubber.add_stub(sp1)
 		gr.stubber.add_stub(sp2)
 
 		var d = autofree(gr.doubler.double(DoubleMe).new())
-		assert_eq(d.has_one_param(), 5)
+		assert_eq(d.return_string_plus_a(), '5')
 		if(is_failing()):
 			print(gr.stubber.to_s())
 
@@ -126,9 +128,9 @@ class TestTheBasics:
 
 	func test_can_stub_native_methods():
 		var d_node2d = autofree(gr.doubler.double_gdnative(Node2D).new())
-		var params = GutUtils.StubParams.new(d_node2d, 'get_position').to_return(-1)
+		var params = GutUtils.StubParams.new(d_node2d, 'get_position').to_return(Vector2(-1, -1))
 		gr.stubber.add_stub(params)
-		assert_eq(d_node2d.get_position(), -1)
+		assert_eq(d_node2d.get_position(), Vector2(-1, -1))
 
 	func test_partial_double_of_Node2D_returns_super_values():
 		var pd_node_2d  = autofree(gr.doubler.partial_double_gdnative(Node2D).new())
@@ -136,9 +138,9 @@ class TestTheBasics:
 
 	func test_can_stub_all_Node2D_doubles():
 		var d_node2d = autofree(gr.doubler.double_gdnative(Node2D).new())
-		var params = GutUtils.StubParams.new(Node2D, 'get_position').to_return(-1)
+		var params = GutUtils.StubParams.new(Node2D, 'get_position').to_return(Vector2(-1, -1))
 		gr.stubber.add_stub(params)
-		assert_eq(d_node2d.get_position(), -1)
+		assert_eq(d_node2d.get_position(), Vector2(-1, -1))
 		if(is_failing()):
 			print("Node2D = ", Node2D)
 			print(gr.stubber.to_s())
@@ -190,8 +192,6 @@ class TestInnerClasses:
 class TestDefaultParameters:
 	extends BaseTest
 
-
-
 	var doubler = null
 	var stubber = null
 
@@ -234,6 +234,22 @@ class TestDefaultParameters:
 		stubber.add_stub(params)
 		assert_eq(stubber.get_default_value(dbl, 'defaulted_second_parameter', 0), null)
 		assert_eq(stubber.get_default_value(dbl, 'defaulted_second_parameter', 1), 'p2')
+
+
+class TestReturnTypes:
+	extends GutInternalTester
+
+	var doubler = null
+	var stubber = null
+
+	func before_each():
+		doubler = GutUtils.Doubler.new(GutUtils.DOUBLE_STRATEGY.INCLUDE_NATIVE)
+		stubber = GutUtils.Stubber.new()
+		doubler.set_stubber(stubber)
+
+
+
+
 
 
 
@@ -281,10 +297,10 @@ class TestMonkeyPatching:
 
 	func test_with_bound_parameters():
 		var dbl = autofree(doubler.double(DoubleMe).new())
-		var params = GutUtils.StubParams.new(dbl.has_two_params_one_default)
+		var params = GutUtils.StubParams.new(dbl.two_params_variant_return)
 		params.to_call(return_passed.bind('three', 'four'))
 		stubber.add_stub(params)
-		var result = dbl.has_two_params_one_default('one', 'two')
+		var result = dbl.two_params_variant_return('one', 'two')
 		assert_eq(result, ["one", "two", "three", "four", null, null])
 
 
@@ -333,7 +349,7 @@ class TestSingletons:
 			.when_passed('foo', true)
 		stubber.add_stub(params)
 		var dbl = doubler.double_singleton(OS).new()
-		assert_eq(dbl.get_system_dir('bar', true), null)
+		assert_eq(dbl.get_system_dir('bar', true), "")
 		assert_eq(dbl.get_system_dir('foo', true), '/asdf')
 
 	func test_can_stub_method_to_return_based_on_parameters_without_specifying_defaults():
@@ -342,7 +358,7 @@ class TestSingletons:
 			.when_passed('foo', true)
 		stubber.add_stub(params)
 		var dbl = doubler.double_singleton(OS).new()
-		assert_eq(dbl.get_system_dir('bar'), null)
+		assert_eq(dbl.get_system_dir('bar'), "")
 		assert_eq(dbl.get_system_dir('foo'), '/asdf')
 
 	func test_can_stub_singleton_to_call_a_func():